Refactor Sqlite database create SQL syntax
The SQL syntax to create the Snippy database was formatted
with [1] and changed to better align with PostgreSQL and
CockroachDB.

There seems to be no effects from the change since all the
tests are passing.

The BLOB was changed to CHAR which both seem to map to TEXT
in Sqlite [1].

Timestamp defaults were removed because they must be always
set in the implementation. The implementation must not rely
on database defaults. The new syntax on timestamp is same
with PostgreSQL. The CockroachDB would require TIMESTAMPTZ
but that is handled in a later commit. Sqlite seems to use
TEXT for datetime and timestamp so this does not impact [1].

[1] https://www.sqlite.org/datatype3.html

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>