Fix importing content without UUID field
If the UUID was missing from imported JSON, YAML or Markdown
content, the empty UUID caused unique constraint violation
from database UUID column. The reason was that two empty
strings are considered same.

If the UUID is empty when the content is parsed or created,
it is now always allocated.

The case is not relevant with text content because text
content input template does not have UUID field.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>