Change externally visible UUID to UUID4 (breaking)
The externally visible UUID field was changed from UUID1
to UUID4. The reason is that the UUID1 contains host MAC
address that should not be visible outside the server.

The externally visible UUID is allocated only once. There
is always a change of UUID4 collision if multiple servers
are run but it is considered so small that it does not
matter [1]. You would have to "generate a billion UUIDs
every second for a century and win a coin flip" to get
UUID4 collision.

There will be internal UUID that will be intended to
be used as a database primary key. This will contain
host namespace or unique identifier that guarantees
the uniqueness well enough within one server.

You need to set the existing content uuid field in
exported YAML file to empty sting '' and then import
the content again in order to change the UUID1 field
to new UUID4 field.

  $ snippy export --all --defaults
  # Manually set all uuid fields to uuid: ''
  $ snippy import --all --defaults

[1] https://en.wikipedia.org/wiki/Universally_unique_identifier#Version_4_(random)

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>