Fix reading timestamps without quotes from YAML
It seems that yaml.safe_load automatically converts unquoted
timestamps to Datetime objects. This caused change in the
timestamp format when the content was imported and exported.

This may have been broken while developing 0.9.0 version and
it is not sure if this exists in version 0.8.0. This may be
because of the UTC timestamp format change from UTC offset
+0000 to +00:00. It may be that the Yaml load did not
recognize the previous format. If this is the case, this is
not affecting version 0.8.0.

The timestamp format follows the ISO869 format in UTC time.
Exporting the data now forces the exported timestamp in YAML
file to be quoted string.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>