Fix SQL database primary key
The primary key was not properly used before. The syntax also
was not compatible between Sqlite, PostgreSQL and CockroachDB.

The primary key is now allocated as UUID1. It still needs to
be checked if containers use same MAC address that is part
of the UUID1. If this [1] is the case, there may be UUID
collision (not very likely) between different servers running
in different containers.

The primary key ID was also placed as a first column which
seems to be common practise. There is controversial discussion
about naming the primary key as ID but it seems to be not
directly against best practise.

The PRIMARY KEY is automatically unique and not null at least
with PostgreSQL and perhaps also CockroachDB which follows
PostgreSQL. But with Sqlite, the primary key allows NULL values
in primary keys that are not INTEGER [2]. Since the primary key
is UUID, there is the 'NOT NULL UNIQUE'.

The metadata column was removed since it was never used. The
original thought was that the metadata could contain bits to
indicate something. But it is better to add new columns if
needed.

[1] https://docs.docker.com/engine/reference/run/#network-settings
[2] http://www.sqlitetutorial.net/sqlite-not-null-constraint/

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>