Change DB schema to be compatible with PostgreSQL
The database schema was changed so that the timestamps are
in format of 'TIMESTAMP WITH TIME ZONE'. This format is
compatible with Sqlite and PostgreSQL. It should be also
compatible with CockroachDB as well.

The timestamps are stored as a strings and the database
timestamp columns are just a place holders for the times.
The timestamps are always stored in UTC offset of +00:00.
The usage of TIMESTAMP WITH TIME ZONE tries to clarify
the usage of UTC based timestamps even though it could
be done without time zone information as well.

By defining the timestamps with time zone aware timestamps,
it should be possible to adjust the database time zone to
get local time stamps if needed (this was not tried).

The external UUID field SQL type was changed to UUID to
reflect that it is an UUID.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>