Change cause when there is no content exported
If a search criteria or specific command did not find any
content to be exported, the exit cause was changed from OK
to NOK. This is more clear to end user because OK would
indicate that there would be for example a file created
when --file option is used.

The cause is now 'NOK: no content found to be exported' in
this specific case.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>