Add command line options to configure database
Command line options were added to configure other database
than Sqlite. The server configuration options were changed
because SSL certificates must be possible to set separately
for database and the server. For example the SSL certificate
Common Name (CN) ties the certificate to specific host and
it can be that server and database are running in different
hosts.

The storage host is set by --storage-host and not with a
separate IP and port parameters. It was considered that
writing something like 'localhost:8080' is easier for user
than writing two different parameters.

Also there is no default for the storage host in order to
enforce user to define them correctly and securely.

Same rules are planned to be enforced to server options.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>