Edit TODO list
It may be that the Sqlite SQL transactions are autocommit and
rollback with with statements (context manager). So it may be
that the 'commit' and 'rollback' statements are not really
needed with INSERT, UPDATE and DELETE. But since there is work
ongoing to use also other databases, it may be better to leave
these statements as of now.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>