Add support to manage database from command line
It is now possible to define PostgreSQL as a server
backend to persist data. The command line option
example is:

  --storage-type postgresql --storage-host localhost:5432 --storage-database postgres --storage-user postgres --storage-password postgres

when running the server. In case the parameters are not
defined, the default Sqlite is used.

Running tests with Pylint option --snippy-db postgresql
always requires PostgreSQL set-up with default settings.
That is, the test cases contain currently hard coded
values for host, database and user credentials.

The SSL parameters are not tested. This is a work in
progress.

Note that that server startup will change in near
commits so that the server requires the server-host
in a same manner as the storage-host.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>