Change cause code message printing
The cause code message is now enforced to be a one liner.
This was the functionality already before when the cause
was set by the Snippy tool. The cause setting is used now
also from exceptions and those may contain more data than
one line. The specific case was changed in previous commit
related to PostgreSQL database connection problem.

The cause message is intended to be a short one liner
for end user. It is considered that used should not get
multi line complex string from exception but a short
meaningful message. For more detailed information and
with full message, user can run the tool with -vv or
--debug options.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>