Fix database exception test with PostgreSQL
The implementation is made so that Sqlite3 is always imported
because it is part of Python distribution. If PostgreSQL is
used, it requires installing Psycopg2 package. It is possible
to run the code without installing Psycopg2 package.

In order to avoid 'undefined' exceptions from Psycopg2 module
when it is not installed, there are dummy definitions from
two exceptions that are used from the Psycopg2 module.

Author was not able to figure out a way to handle this with
one custom exception. One tried alternative was to import
the IntegrityError from both packages with 'from x import
x.IntegrityError as IntegrityError'. This was failing when
both packages were loaded since the later Psycopg2 override
the IntegrityError which caused faults when Sqlite3 was used.

There was also no way found out to inherit and mask both of
the exceptions behind single custom exception. In this case
the new custom exception was not caught. The problem was
likely that Sqlite3 and Psycopg2 raise their exceptions
and not the new custom exception.

The Exceptions tests can be done by throwing Sqlite3 Exception
like in test test_api_create_reference_008 when testing. This
works because the Sqlite is always bundled with the release
and. The testing should be possible also with Psycop2
exceptions but that is not tried at the moment.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>