Change content search behavior with UUID field
The UUID is unique field between all the content. The UUID
field is defined as UUID in the SQL schema. Sqlite3 maps
an UUID field to text and it supports SELECT queries like
'WHERE LIKE'. In case of PostgreSQL, selecting like this
from unique UUID field is not possible. This is likely
related to optimization where unique fields cannot result
more than one row.

Because of this, the UUID search behaviour was changed.
The change is that now the content UUID must be always
fully defined to the CLI or API. If the UUID is partial,
it does not match any content.

The content digest can be used to search multiple contents
with a short digest.

The UUID field is new to development release 0.9.0 and
thus this does not affect as a change from version 0.8.0.

It may be considered that the REST API can be operated
with digest or UUID. Both of them offer a bit different
characteristics for different purposes.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>