Fix DataError exception with external UUID field
If UUID field is not a correctly defined UUID, the PostgreSQL
database will throw DataError Exception. Previously this was
handled only as generic Error which created Internal Server
Error.

Because the error is generated by external data, it was
considered that the Internal Server Error is not good here.

The problem is that Sqlite and PostgreSQL behave differently
here. There is no such exception in this case from Sqlite.
Because the end user behavior is wanted to keep the same
also with Cause codes, the DataError is captured but it
does not generate Cause. There are Cause codes set from
other locations that will reveal the problem for end user.

This problem applies to externally visible content UUID
field. The internal UUID field is always assumed to be
correctly set since that must be always validated. If
the internal use of internal UUID field fails or there
is an internal case to use the UUID, failures from these
generate Internal Server Error.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>