Change content operation output to be deterministic
When the content was searched, sorting of the content was
made based on created timestamp. When the content was
exported, the order was not sorted and it was the order
how database query resulted the data.

Sorting of database query responses is different between
Sqlite3 and PostgreSQL. This revealed problem with tests
because the API result and test case assert use a list
with dictionary based contents. This enforces specific
order for the data which was now different with new
database.

By default if user does not define otherwise, the sort
for search and export operations is now based first
on created timestamp in ascending order and then based
on brief field.

The same applies for the tool and database helpers for
testing.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>