Fix Unicode conversion with Python 2 and PostgreSQL
In case of Python 3, the Unicode conversion is made
automatically. With Python 2 specific extension must
be used to get Unicode by default [1].

The UTF-8 Unicode is always used internally and the
external sources are always converted to UTF-8
Unicode strings.

The PostgreSQL settings now define the locally used
connection as UTF-8 which uses the new Unicode
extensions.

All tests now pass also with Python 2 with Sqlite3
and PostgreSQL.

[1] https://stackoverflow.com/a/14887474

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>