Add PostgreSQL Psycopg2 adapter as server dependency
The setup.py, requirements and virtual environment files
now include the Psycopg2 binary dependency. In live usage
the Psycopg2 is recommended to be compiled for target
environment.

It was also considered that there would be own 'server-
postgresql' install target. But this was combined with
the server target as of now. The reason was that Pipenv
has only one 'dev' alternative for installing development
dependencies (as far as known).

There is room for improvements related to installing with
Sqlite or PostgreSQL and choosing correct dependencies.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>