Change server command line options (breaking)
The server command line options were completely changed. This
aligns the server options with storage options.

There is no default value anymore for the server bind IP and
port. This tries to enforce user to think which IP to bind
and to avoid leaving unsecure configurations running. User
must always define '--server-host' parameter in format of
<ip>:<port>. For example --server-host 127.0.0.1:8080 is
valid configuration.

There are no checks to parse the IP address or possible
DNS name. This would be complex and error prone so it is
not done as of now. Instead the server startup will just
fail if the port is reserved or the address or host is
not valid. Currently this happens silently which should
be changed to give proper error cause for user.

Changed the --compact-json flag to --server-minify-json.
Also internally all the server variables were renamed to
have prefix 'server' like with the storage parameters.
This naming rule helps grouping of different parameters
and to read the code.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>