Change default file format to Markdown
When content is imported or exported, the default file
format is now Markdown. This does not apply to default
content that is always operated in YAML format.

This tries to make  Markdown format as the main format
that is used to operate the tool.

The proprietary text format is not going to be removed.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>