Change Markdown formatted solution parsing
The Markdown formatted solution must start with a second level
Markdown header and it is ended with the Meta header. The second
level Meta header and the content is automatically generated by
the tool.

The reason to use second level headers to indicate a Markdown
formatted content is that the Solution header part is generated
by the tool as fist level header. The intention is that all
the solution content will fall under the first level header.
This allows printing multiple Markdown formatted contents
into same file and they are nicely separated with a first
level header.

This also fixes case where pure Markdown formatted content
was exported into Markdown file when the Markdown formatted
solution was placed inside Markdown code block. This caused
the actual content to render incorrectly. The code block
syntax is needed only when the solution is not Markdown
native content.

This change targets to make the Markdown as first choice
to operate the content from command line.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>