Fix updating Markdown native solution in editor
If Markdown native solution is updated, the editor will
now show the Markdown content. The Markdown content is
presented with header and metadata parts which must
retain their specific formats. If the header part or
metadata parts are updated with incorrect format, the
update will fail.

The header part starts from the beginning of the file
and ends with the first second level Markdown header.
The metadata part starts with the second level Meta
header.

The links in the header part are updated automatically
from the solution data that follows after the header.
The first line '# Testing docker log drivers @docker'
contains the brief and groups which can be modified
to update reflected values in the content.

The Metadata contains fields 'digest' and 'updated'
which are always updated by the tool regardless of
the user input.

Example below shows editor view:

  # Testing docker log drivers @docker',

  > Investigate docker log drivers and the logs2kafka log plugin

  > [1] https://github.com/MickayG/moby-kafka-logdriver
  [2] https://github.com/garo/logs2kafka
  [3] https://groups.google.com/forum/#!topic/kubernetes-users/iLDsG85exRQ

  ## Description

  Investigate docker log drivers.

  ## Solutions

  ## Whiteboard

  ## Meta

  > category : solution
  created  : 2019-01-04T10:54:49.265512+00:00
  digest   : 18473ec207798670c302fb711a40df6555e8973e26481e4cd6b2ed205f5e633c
  filename : kubernetes-docker-log-driver-kafka.mkdn
  name     :
  source   :
  tags     : docker,driver,kafka,kubernetes,logging,logs2kafka,moby,plugin
  updated  : 2019-01-05T10:54:49.265512+00:00
  uuid     : 24cd5827-b6ef-4067-b5ac-3ceac07dde9f
  versions :

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>