Fix updating Markdown formatted solution links
The links in the Markdown solution header are updated
automatically. The links are read from the solution
content data. Previously the Markdown parser read only
the header links and ignored the links in the content
data.

This applies for solution content only. For the snippets
and references, links are read from the Markdown header
part.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>