Change solution Markdown template
The solution content Markdown template now contains the
default header structure. This tries to help and speed
up adding the content when a new solution is created.

This is added programmatically on top of the default
Markdown template in order to avoid changing the default
template. The default template is intended to be allow
easy presentation of the content in Markdown format. This
should make the code easier to to read.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>