Fix snippet Markdown template
There were unnecessary new lines when new snippet was created
from editor. These are now removed. The newlines were coming
from the <data> tag replace which was just replaced with an
empty string which left two unnecessary new lines for empty
snippet template.

There was no nice way figured out to remove this. Because of
this, there is a specific condition for 'snippet and template
content'. That is the removal is specific for this which is
not too generic. This should have been in the Resource()
get_template method but as above, there was no clean way
figured out to do this.

Added also tests to create Markdown content from editor for
all content categories.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>