Change editor option and add no-editor option
The editor is by default the preferred way to create and
update content. The reason is that it is harder to write
the data to command line since the all content categories
tend to be quite long.

The editor is used always with create and update operations
if user did not provide mandatory links for reference content
and data for other content categories.

Because the editor is more often used now, there is now
'no-editor' option that can prevent the usage of editor. The
'editor' option works the same as before and it enforces the
editor usage. The 'no-editor' has precedence over 'editor'
option.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>