Change cause setting in case of empty template
The Resource tests for empty template or empty data previously
set Cause. This was problematic when these methods were just
needed to test these two conditions because the Cause was sent
to end user in REST or or it could affect unintentionally to
CLI result.

Now the cause is set by the caller of these two methods when
needed.

The order was changed also so that the test for empty template
has higher precedence over empty data. It is considered more
self evident if used does not change anything and cause with
text referring to empty template than just one empty field
'data'.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>