Add help text when creating new Markdown content
When new content is created in Markdown format from command
line, a help texts are added on top of the empty template.
This tries to help end user to fill the template.

This is not considered perfect. For example the groups field
is not clearly indicating that comma separated values are
supported. Also the list of http links may not be self
evident. The thought is that the user is somewhat familiar
and is able to use examples. If there is too many help texts
added, it is considered slowing down more experienced user.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>