Fix Markdown export format for snippet data
The Markdown syntax of two spaces at the end of line enforces
newline. This was missing when snippet data was exported.

There is now explicit test that verifies this. This is not
noticed by the used testing framework. This is calculated
drawback that is explained in more detail in 'Using collections
for asserting expected content' section of document scribling
document.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>