Refactor Resource digest computation
The digest is considered as internal parameter for Resource
so that it is always set by the class. This change removed
unnecessary digest computations outside the Resource. The
Resource seal() method always guarantees that that this
attribute is set correctly.

This could have also very small performance impact because
the digest is not computed unnecessarily anymore when a
collection is formed from parsed text file or API call.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>