Test parsing repeated blockquote in Markdown links
There are two formats for Markdown user input for link list
are supported. First one, which is also used by the tool
when exporting or printing the data is a single blockquote
in the first item like:

> [1] https://docs.docker.com/engine/reference/commandline/images/
[2] https://docs.docker.com/engine/reference/commandline/rm/
[3] https://docs.docker.com/engine/reference/commandline/rmi/

The second format is with blockquote in the beginning of
each line like:

> [1] https://docs.docker.com/engine/reference/commandline/images/
> [2] https://docs.docker.com/engine/reference/commandline/rm/
> [3] https://docs.docker.com/engine/reference/commandline/rmi/

It was considered back and forth which would be cleaner
default format in export and the first example was chosen.
The main reason was to try to reduce glutter in the output.
This is not self evident by default for end user so it was
made sure that the parser supports both formats when a
Markdown content is read.

The same reasonings apply also to the metd chapter. There
it is assumed that user does not modify the metadata keys,
only the values (of selected keys).

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>