Change cause code message print
Unnecessary column delimiters were removed from cause
code message.

Cause codes follow the same rules as the logs with the title or
message. Also the same message that goes to cause message goes
automatically to log. If there are variables within the message,
the variables are separated with colon. End user message is
beautified so that if there is more than one colon, it indicates
that variable is in the middle of the message. This is not
considered good layout for command line interface messages and
thus the colons are are removed.

How ever, if there is only one colon, it is used to separate
the last part which is considered clear for user.

Because of these rules, the colon delimiters are removed only if
there is more than one.

Examples where the first message is manipulated and the second
remains as is:

  1. cannot use empty content uuid for: delete :operation
  2. cannot find content with content uuid: 1234567

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>