Add support to give content fields for update
If user updated content from command line and gave some
of the content fields like links or tags, the additional
parameters were not used. If user runs update operation,
the additional parameters are used to update the updated
content.

For example if editor is used to update existing content,
user can now give --link option which value is updated
automatically on top of existing content.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>