Fix resource validation for PUT and PATCH methods
When content is created, there has to be a data or links
fields depending on content type. But in case of updating
existing content, it is possible to leave the mandatory
content field out.

In case of a REST API server, the PUT or PATCH method
JSON structure was incorrectly validated to require
always the data or links field. This is now fixed so
that a POST always requires this but other methods
allow update without mandatory fields.

A minor refactoring made to always call the Validate
class json_object method for REST API method handlers.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>