Fix exception in server if content not found
If there was a special error case where REST API server
received a request that did not result any content in
response, there was chance of internal error condition
would trigger another error which would generate
exception from variable that was not defined.

This condition would require internal error to store
content that would not have mandatory content data
or links depending on content type. This condition
when sending a response to REST API client could
generate failure to read internal undefined variable.

This is in category "never should happen" which are
always occurring in live cases.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>