Fix resource validation for PUT method
This was an error caused from the latest change with
commit message [1]. Because PUT request replaces all
content field, the PUT request is also considered to
create a new content. In case of creating new content
or replacing all the fields, the mandatory fields must
be present.

Previously this change was made only for POST.

[1] Fix resource validation for PUT and PATCH methods

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>