Fix PATCH content attributes to empty values
There was an error where a client was not able to clear values
of optional content fields like 'brief' or 'tags' with PATCH
method.

Now the REST API works according to JSON Merge Patch - RFC 7386.
If a value is requested to be cleared with with a PATCH, a null
value must be used. The attribute is not deleted, it is just
set to default empty value.

With the same commit, the Parser class was improved to handle
None values with strings and lists. The Resource class also
uses now the Parser class to set values for Resource attributes.
This allows directly setting the Resource attribute to None and
the Parser returns the correct default value of empty tuple or
an empty string.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>