Change content field versions to array (breaking)
The versions field was changed from string to array. It is
expected that content may have for example multiple third
party software projects referred in which case there is a
need to have multiple versions.

Because this changes the internal behaviour of the field,
you must export the old content and import it again with
new version of a tool.

  $ snippy export --defaults --all
  $ snippy import --defaults --all

Another option is to manually change the versions fields
in source Yaml from string to array.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>