Remove merge option from command line
The merge option did not have previously any relevance
from command line. The option was not documented in the
tool help text. Because of these, the change is not set
breaking change.

The CLI behaviour is always merge. This allows updating
existing content from CLI so that the empty values for
CLI fields that are not defined, do not affect to updated
content.

The merge is meant to be used from API where it allows
differentiation between PUT and PATCH. The PATCH is
internally defined as 'merge' set to True.

This allows also cleanup of internal implementation.
The content base.py does not have to care if the content
is merged or migrated because the Config() is always
making this decision always from Config.merge option.

There is now tested support to give parameters from
command line to override existing content. For example
below command adds a tag from command line to editor
which shows the existing content and the tag when the
content is presented for user in editor.

  $ snippy update -d f3fd167c64b6f97e --tags new,tags,from,cli

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>