Fix exception handling when deleting resource
When resource was not successfully migrated to collection,
it's delete might have caused exception which was not
captured causing exception and halt of the program.

This is now taken care as per rule that there are no
exceptions thrown to users. All exist are made gracefully
shutting services (when necessary) and showing proper exit
cause to user.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>