Refactor parsing snippet data from Markdown
There was an effort to loosen the requirements of putting
snippet commands between backtics with leading dollar sign.
This failed because it was not possible to loosen parsing
conditions since a command may contain any quotation marks
or backtics. The only safer way to parse commands from
Markdown was to force user to give the commands with `$ `.

Clarified code comments regarding of the above.

Refactored snippet data parsing by storing the parsed
snippet in tuple instead of a string. It is more logical
to store parsed commands in a list instead of a single
string where commands were separated by newlines.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>