Change content default format to Markdown
Markdown format is now the default format for command line
operations. Only when content is printed on terminal as a
result of search operation, the default is tool proprietary
text format.

It is possible to control the format with `--format` option
from command line. For example, it is possible to use text
templates when creating content with examples:

  $ snippy create --format text
  $ snippy create --reference --format text

It is also possible to output the search result directly in
Markdown format with `--format` option like:

  $ snippy search --sall tar --format mkdn

The default format for import and export operation is also
Markdown. If export or import operation does not define a
filename, the default file is in Markdown format. It is
possible to export and import Yaml, Json or Text files by
defining *.yaml, *.json or *.text file extension for the
`--file` command line option.

This change tries to make the Markdown format as the main
format in which end user sees the content. Internally
content is always in dictionary format because it can be
converted to any presentation format.

This change does not affect to REST API server and the REST
API server operates with JSON data before.

This change converted the default solutions to native Markdown
format.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>