Fix Dockerfile for PostgreSQL psycopg2 adapter
The psycopg2 PostgreSQL adapter must be compiled for Alpine
because it uses MUSL. The Alpine container with MUSL is not
compatible with psycopg2 binaries.

Also the libpq is needed by the PostgreSQL in the container.

The image size grew from 32MB to 36.8MB which is considered
pretty OK.

There are cleanup to be done for the Dockerfile:

  - Use Python 3.7
  - Copying psycopg2 files is version specific from base.
  - Python version is assumed in commands to be 3.6.
  - There is a hack to rename the psycopg2 to psycopg_binary.
  - Check if there is room for image size optimization.

Installing the software with psycopg2 (not the binary) was
not working out of the box so the hack to rename the file
was left as of now.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>