Fix warning in Docker compilation
The APK installer seems to ignore --no-cache and create cache
files anyways [1]. Removing the --no-cache removes the below
warning. This did not increase the Docker image size.

  $ make docker
    ...
    Uninstalling pip-19.0.1:
      Successfully uninstalled pip-19.0.1
    WARNING: Ignoring APKINDEX.adfa7ceb.tar.gz: No such file or directory
    WARNING: Ignoring APKINDEX.efaa1f73.tar.gz: No such file or directory
    (1/1) Purging apk-tools (2.10.1-r0)
    ...

Removed also Python distutils which reduced the image size
with 0.8 MB. The distutils is related to Python installer
which is not needed since the whole pip and Python installing
is intended to be prevented in Docker image.

Minor refactoring to sort the installed packages on alphabetic
order.

[1] https://github.com/gliderlabs/docker-alpine/issues/207

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>