Fix error when content matching to a template is saved
If snippet, solution or reference text template was
saved from text editor, the error string complained
about missing data field. It is considered that when
content is matching to a template, it is greater error
and thus it should be reported instead of missing data.

Corrected error text when saving text template from
editor without any changes as in three examples below:

  $snippy create  # Launch vi editor.
  :wq

  $ snippy create --solution
  :wq

  $ snippy create --reference
  :wq

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>