Fix exporting and importing content template
There was a problem that the 'export --template' still
exported text template by default. The default must be
Markdown formatted template.

The --template option now follows the --format option
with default value set to Markdown.

When importing the template, the format must be set
with the --format option if --template option is used
as a shortcut. The --file option it is possible to
automatically read the content format and there is no
need to use --format option.

The content options --snippet, --solution or
--reference can be used to select the content template
category.

  $ snippy export --template  # Export md template.
  $ snippy import --template  # Import md template.

  $ snippy export --template --text  # Export text template.
  $ snippy import --template --text  # Import text template.

  $ snippy export --template --text
  $ snippy import --file ./snippet-template.text

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>