Refactor Dockerfile and update Alpine to 3.9
There is now own install target for Dockerfile. The
psycopg2-binary compilation had potential problems
because it was using hard coded version numbers for
Python and for the psycopg2.

The Dockerfile is now more generic because there is
an own setup target for Docker. This allows using
the psycopg2 which can be installed into Alpine
with the apk installer. How ever, the Alpine does
not have the latest psycopg2 thus the version
settings differs for the server and Dockerfile.

This is not perfect since there are now slight
differences of versions and modules between
server and docker setup targets. But this now
cleans the Dockerfile and makes it more
maintainable.

Updated Alpine to latest 3.9 version.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>