Fix make server target
There was previous correction that incorrectly added the
'pip install -e .' into server target. This is not needed
because the server is an extra target in setup.py on top
of the install_requires.

This means that when the make server is run, all required
modules including pyyaml are installed.

This applied also to Dockerfile which was updated in the
previous commit.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>