Fix Dockerfile psycopg2 installation
There is now maximum verision for psycopg2 installed into
Docker image. The limit is the same as the pinned version
of the server extras in the setup.py. The server is the
main development extras to develop the server. This tries
to prevent accidentally using later version in Docker
image.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>