Add documented workaround for piping with more
If ANSI color codes are used when the command line output
is piped to the more command, there are extra 'e--' marks
in some of the line ends.

These are left from --More-- and they are not coming from
the tool. It is not know if this a bug in the Snippy tool,
expected behaviour in more or a bug in the more command.

The lines that are empty likes with '   : ' with ANSI
color codes which are shorter than the '--More--' string,
leave tail 'e--' from the more command. This is likely
somehow related to color codes and character string lengths
somewhere.

It is possible to use more options -c or -p to workaround
this issue. This is visible at least with Fedora 26 with
more from util-linux 2.29.1.

Example command to produce:

  $ snippy search --sall elastic --solutions | more

Example commands to workaround:

  $ snippy search --sall elastic --solutions | more -c
  $ snippy search --sall elastic --solutions | more -p

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>