Fix broken pipe error with grep and tool logs
Also code refactoring to move the print to stdout
to Logger to have common place to run the code with
SIGPIPE signal manipulation.

If the logs are enabled with -vv or --debug option and
there is a broken pipe error with grep, logs generated
exceptions with example command below.

  $ python runner search --sall 'test' --filter test -vv | grep --all

The solution is to disable exceptions from the logging
module. The logs cannot be printed to stdout because
it is not available anymore. There was an effort to
override logging StreamHandler class with custom
ErrorHandler. The intention was to write the logs into
stderr when there is a 'broken pipe' error. The problem
is that this exception is not in Python 2 and thus making
generic code is not easy. It was considered too large
effort for the benefit of getting logs when there is a
failure with grep.

There is still an exception below with Python 3:

  $ python runner search --sall 'test' --filter test -vv | grep --all
    grep: unrecognized option '--all'
    Usage: grep [OPTION]... PATTERN [FILE]...
    Try 'grep --help' for more information.
    Exception ignored in: <_io.TextIOWrapper name='<stdout>' mode='w' encoding='UTF-8'>
    BrokenPipeError: [Errno 32] Broken pipe

This is not there with Python 2. There is no know
solution to prevent this at the moment. This seems to
be related to [1] and [2].

[1] https://stackoverflow.com/a/18954489
[2] https://bugs.python.org/issue11380

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>